secondary/pyproject.toml
[project]
name = "secondary"
version = "0.1.0"
description = "Implementation of replicated log: secondary node"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "uvicorn>=0.31.0",
]
